+++ src/WorkOS/Options/Production.php
@@ -51,7 +51,6 @@ class Production extends Options {
 			'change_email_rate_limit_ip_window'       => 3600,
 			'change_email_notify_old_address'         => true,
 			'change_email_require_reauth'             => true,
-			'change_email_admin_bypass_verification'  => false,
 			'change_email_confirm_path'               => 'workos/change-email',
 			'allow_magic_code_registration'           => true,
 			'allow_legacy_magic_code_registration'    => true,